🍳 Rename "egg" emoji to "cooking"
The unicode spec uses "cooking" as the name for emoji U+1F373 (🍳), and
it uses "egg" as the name for emoji U+1F95A (🥚).

  http://emojipedia.org/emoji/%F0%9F%8D%B3/
  http://emojipedia.org/emoji/%F0%9F%A5%9A/

This commit renames 🍳 to "cooking" to match the unicode spec and to
make room for adding 🥚 and giving it the name "egg" in a future commit.